modify for support python3, python2 old-fashioned

+++ shooting_plane/game.py
@@ -206,10 +206,10 @@ while running:
     if not player.is_hit:
         screen.blit(player.image[player.img_index], player.rect)
         # 更换图片索引使飞机有动画效果
-        player.img_index = shoot_frequency / 8
+        player.img_index = int(shoot_frequency / 8)
     else:
         # 玩家飞机被击中后的效果处理
-        player.img_index = player_down_index / 8
+        player.img_index = int(player_down_index / 8)
         screen.blit(player.image[player.img_index], player.rect)
         player_down_index += 1
         if player_down_index > 47:
@@ -232,7 +232,7 @@ while running:
             elif score <= 200000:
                 SCORE_FREQ = 20
             continue
-        screen.blit(enemy_down.down_imgs[enemy_down.down_index / 2], enemy_down.rect)
+        screen.blit(enemy_down.down_imgs[int(enemy_down.down_index / 2)], enemy_down.rect)
         enemy_down.down_index += 1
 
     # 显示子弹
